Add files via upload
Introduction
By now, you have explored a range of functions that you can use to generate date and time results in your spreadsheets. 

In this exercise, you’ll use this knowledge to generate date information and date calculations to help with the development of a project plan. You’ll use the functions TODAY, NETWORKDAYS, MONTH, and YEAR to add some important milestone information to a spreadsheet.

Case study
Adventure Works is preparing a series of advertising campaigns to be rolled out in several different regions. A colleague, Lucas, has asked you to update a spreadsheet that focuses on the launch dates for the USA campaign. The spreadsheet is called Advertising Campaign USA Dates.xlsx.

For each project, Lucas needs to know the following information:

The number of working days available between the start date and the deadline date.

The month and year when each campaign will launch.

The number of calendar days to the deadline date for each campaign.

Let’s begin creating the formulas that will generate this information for Lucas.

Step 1: Download the file.
Download and open the Microsoft Excel workbook Advertising Campaign USA Dates.xlsx. The workbook contains only one worksheet called USA Launch Dates.

The sheet contains information about the five USA campaigns, including the start date and launch date for each campaign. It also includes a list of the dates of Federal holidays for both 2023 and 2024 and the current date in cell B1. The date in cell B1 is to be used as the "start date" in the calculations you create in E5 and F5. 

Important: Please note that the Excel shown in the screenshots in this exemplar is set to United States settings and dates are therefore showing in the MM/DD/YY format. If your local machine is set to a different region and does not display dates in this format, please skip Steps 1 and 2 in this exercise. Completing them may cause the format setting on cell B1 to revert to your local region's date format. The calculations you create would then produce results different to those shown in the screenshots. 

Conclusion
This spreadsheet provides a good example of how date functions can be used to generate new date and time information in a spreadsheet. 

In this exercise, you used a selection of date functions and calculations to generate new milestone information in the data. You used the correct syntax for each formula and customized the cell references so that the formulas could be copied down. This action is a huge time-saver. 

Lucas can now use the details in the spreadsheet to begin planning the workflow for the rollout of the campaign.